Add error handling on instruction_decoding
