+++ src/vm/decoding/decoder.rs
@@ -1,4 +1,5 @@
 use crate::types::instruction;
+use crate::vm::vm_errors::VirtualMachineError;
 use num_bigint::BigInt;
 use num_traits::FromPrimitive;
 
@@ -7,7 +8,7 @@ use num_traits::FromPrimitive;
 
 #[allow(dead_code)]
 /// Decodes an instruction. The encoding is little endian, so flags go from bit 63 to 48.
-pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> instruction::Instruction {
+pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> Result<instruction::Instruction, VirtualMachineError> {
     const DST_REG_MASK: i64 = 0x0001;
     const DST_REG_OFF: i64 = 0;
     const OP0_REG_MASK: i64 = 0x0002;
@@ -50,13 +51,13 @@ pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> instru
     let dst_register = match dst_reg_num {
         0 => instruction::Register::AP,
         1 => instruction::Register::FP,
-        _ => panic!("Invalid instruction"),
+        _ => return Err(VirtualMachineError::InvalidDstReg(dst_reg_num))
     };
 
     let op0_register = match op0_reg_num {
         0 => instruction::Register::AP,
         1 => instruction::Register::FP,
-        _ => panic!("Invalid instruction"),
+        _ => return Err(VirtualMachineError::InvalidOp0Reg(dst_reg_num))
     };
 
     let op1_addr = match op1_src_num {
@@ -64,7 +65,7 @@ pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> instru
         1 => instruction::Op1Addr::Imm,
         2 => instruction::Op1Addr::FP,
         4 => instruction::Op1Addr::AP,
-        _ => panic!("Invalid instruction"),
+        _ => return Err(VirtualMachineError::InvalidOp1Reg(op1_src_num))
     };
 
     if op1_addr == instruction::Op1Addr::Imm {
@@ -81,7 +82,7 @@ pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> instru
         1 => instruction::PcUpdate::Jump,
         2 => instruction::PcUpdate::JumpRel,
         4 => instruction::PcUpdate::Jnz,
-        _ => panic!("Invalid instruction"),
+        _ => return Err(VirtualMachineError::InvalidPcUpdate(pc_update_num))
     };
 
     let res = match res_logic_num {
@@ -89,7 +90,7 @@ pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> instru
         0 => instruction::Res::Op1,
         1 => instruction::Res::Add,
         2 => instruction::Res::Mul,
-        _ => panic!("Invalid instruction"),
+        _ => return Err(VirtualMachineError::InvalidRes(res_logic_num))
     };
 
     let opcode = match opcode_num {
@@ -97,7 +98,7 @@ pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> instru
         1 => instruction::Opcode::Call,
         2 => instruction::Opcode::Ret,
         4 => instruction::Opcode::AssertEq,
-        _ => panic!("Invalid instruction"),
+        _ => return Err(VirtualMachineError::InvalidOpcode(opcode_num))
     };
 
     let ap_update = match ap_update_num {
@@ -105,7 +106,7 @@ pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> instru
         0 => instruction::ApUpdate::Regular,
         1 => instruction::ApUpdate::Add,
         2 => instruction::ApUpdate::Add1,
-        _ => panic!("Invalid instruction"),
+        _ => return Err(VirtualMachineError::InvalidApUpdate(ap_update_num))
     };
 
     let fp_update = match opcode {
@@ -114,21 +115,23 @@ pub fn decode_instruction(encoded_instr: i64, mut imm: Option<BigInt>) -> instru
         _ => instruction::FpUpdate::Regular,
     };
 
-    instruction::Instruction {
-        // TODO: Replace or confirm the unrwap is safe
-        off0: BigInt::from_i64(off0).unwrap(),
-        off1: BigInt::from_i64(off1).unwrap(),
-        off2: BigInt::from_i64(off2).unwrap(),
-        imm,
-        dst_register,
-        op0_register,
-        op1_addr,
-        res,
-        pc_update,
-        ap_update,
-        fp_update,
-        opcode,
-    }
+    Ok(
+        instruction::Instruction {
+            // TODO: Replace or confirm the unrwap is safe
+            off0: BigInt::from_i64(off0).unwrap(),
+            off1: BigInt::from_i64(off1).unwrap(),
+            off2: BigInt::from_i64(off2).unwrap(),
+            imm,
+            dst_register,
+            op0_register,
+            op1_addr,
+            res,
+            pc_update,
+            ap_update,
+            fp_update,
+            opcode,
+        }
+    )
 }
 
 #[allow(dead_code)]
@@ -273,3 +276,4 @@ mod decoder_test {
         assert_eq!(inst.off2, bigint!(1));
     }
 }
+
